+++ changelog.md
@@ -6,6 +6,8 @@
 This version is not released yet and is under active development.
 ```
 
+- Add strict mode to fail on unrecognized configuration options.
+
 ## [1.9.0 (2022-04-08)](https://github.com/kdeldycke/click-extra/compare/v1.8.0...v1.9.0)
 
 - Add supports for `.ini` configuration files.